move serialize and deserialize functions to internal
