+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/key_vault_key.hpp
@@ -14,10 +14,9 @@
 #include "azure/keyvault/keys/key_properties.hpp"
 
 #include <azure/core/http/http.hpp>
-#include <azure/core/internal/json/json.hpp>
 
+#include <string>
 #include <vector>
-
 namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
 
   /**
@@ -79,55 +78,4 @@ namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
      */
     std::vector<KeyOperation> const& KeyOperations() const { return Key.KeyOperations(); }
   };
-
-  /***********************  Deserializer / Serializer ******************************/
-  namespace _detail {
-    struct KeyVaultKeySerializer
-    {
-      // Creates a new key based on a name and an http raw response.
-      static KeyVaultKey KeyVaultKeyDeserialize(
-          std::string const& name,
-          Azure::Core::Http::RawResponse const& rawResponse);
-
-      // Create from http raw response only.
-      static KeyVaultKey KeyVaultKeyDeserialize(Azure::Core::Http::RawResponse const& rawResponse);
-
-      // Updates a Key based on an Http raw response.
-      static void KeyVaultKeyDeserialize(
-          KeyVaultKey& key,
-          Azure::Core::Http::RawResponse const& rawResponse);
-
-      // Create from json node directly. Used from listKeys
-      static void KeyVaultKeyDeserialize(
-          KeyVaultKey& key,
-          Azure::Core::Json::_internal::json const& json);
-
-      void static inline ParseKeyUrl(KeyProperties& keyProperties, std::string const& url)
-      {
-        Azure::Core::Url kid(url);
-        keyProperties.Id = url;
-        keyProperties.VaultUrl = kid.GetUrlAuthorityWithScheme();
-        auto const& path = kid.GetPath();
-        //  path is in the form of `verb/keyName{/keyVersion}`
-        auto const separatorChar = '/';
-        auto pathEnd = path.end();
-        auto start = path.begin();
-        start = std::find(start, pathEnd, separatorChar);
-        start += 1;
-        auto separator = std::find(start, pathEnd, separatorChar);
-        if (separator != pathEnd)
-        {
-          keyProperties.Name = std::string(start, separator);
-          start = separator + 1;
-          keyProperties.Version = std::string(start, pathEnd);
-        }
-        else
-        {
-          // Nothing but the name+
-          keyProperties.Name = std::string(start, pathEnd);
-        }
-      }
-    };
-  } // namespace _detail
-
 }}}} // namespace Azure::Security::KeyVault::Keys
