+++ sdk/keyvault/azure-security-keyvault-keys/src/import_key_options.cpp
@@ -7,6 +7,7 @@
 #include <azure/keyvault/common/internal/unix_time_helper.hpp>
 
 #include "azure/keyvault/keys/details/key_constants.hpp"
+#include "azure/keyvault/keys/details/key_serializers.hpp"
 #include "azure/keyvault/keys/import_key_options.hpp"
 
 #include <string>
